+++ extensions/api/data-management/contractdefinition/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/contractdefinition/model/ContractDefinitionDtoValidationTest.java
@@ -35,7 +35,6 @@ class ContractDefinitionDtoValidationTest {
 
     private Validator validator;
 
-
     @BeforeEach
     void setUp() {
         try (ValidatorFactory factory = Validation.buildDefaultValidatorFactory()) {