+++ extensions/http-provisioner/src/test/java/org/eclipse/dataspaceconnector/transfer/provision/http/webhook/HttpProvisionerWebhookApiControllerIntegrationTest.java
@@ -26,9 +26,11 @@ import org.eclipse.dataspaceconnector.transfer.provision.http.HttpWebhookExtensi
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ArgumentsProvider;
+import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import java.util.Map;
 import java.util.stream.Stream;
@@ -49,20 +51,6 @@ class HttpProvisionerWebhookApiControllerIntegrationTest {
     private final int port = getFreePort();
     private final String authKey = "123456";
 
-    public static Stream<Arguments> invalidRequestParams() {
-        return Stream.of(
-                Arguments.of(null, dataAddress(), "resourcename", "resourcedef", "token"),
-                Arguments.of("assetid", null, "resourcename", "resourcedef", "token"),
-                Arguments.of("assetid", dataAddress(), null, "resourcedef", "token"),
-                Arguments.of("assetid", dataAddress(), "resourcename", null, "token"),
-                Arguments.of("assetid", dataAddress(), "resourcename", "resourcedef", null)
-        );
-    }
-
-    private static DataAddress dataAddress() {
-        return DataAddress.Builder.newInstance().type("foo").build();
-    }
-
     @BeforeEach
     void setUp(EdcExtension extension) {
         extension.registerSystemExtension(ServiceExtension.class, new DummyAuthenticationExtension());
@@ -76,7 +64,7 @@ class HttpProvisionerWebhookApiControllerIntegrationTest {
     }
 
     @ParameterizedTest
-    @MethodSource("invalidRequestParams")
+    @ArgumentsSource(InvalidRequestParams.class)
     void callProvisionWebhook_invalidBody(String assetId, DataAddress cda, String resName, String resDefId, String token) {
         var tpId = "tpId";
         var rq = ProvisionerWebhookRequest.Builder.newInstance()
@@ -140,6 +128,28 @@ class HttpProvisionerWebhookApiControllerIntegrationTest {
                 .then()
                 .statusCode(allOf(greaterThanOrEqualTo(200), lessThan(300)))
                 .body(anything());
+    }
+
+    private static class InvalidRequestParams implements ArgumentsProvider {
+
+        @Override
+        public Stream<? extends Arguments> provideArguments(ExtensionContext context) throws Exception {
+            return Stream.of(
+                    Arguments.of(null, dataAddress(), "resourcename", "resourcedef", "token"),
+                    Arguments.of("assetid", null, "resourcename", "resourcedef", "token"),
+                    Arguments.of("assetid", dataAddress(), null, "resourcedef", "token"),
+                    Arguments.of("assetid", dataAddress(), "resourcename", null, "token"),
+                    Arguments.of("assetid", dataAddress(), "resourcename", "resourcedef", null)
+            );
+        }
+
+        private DataAddress dataAddress() {
+            return DataAddress.Builder.newInstance().type("foo").build();
+        }
+    }
+
+    private DataAddress dataAddress() {
+        return DataAddress.Builder.newInstance().type("foo").build();
     }
 
     private RequestSpecification baseRequest() {